HUGE UPDATE -swiched to component system -object creation is changed -objects behaviours are changed -now we will add the custom component we want to use and that way object will change it's behaviour -rendering changed shape to mesh, system is the same -some error checks added to functions like remove and add. some of them are stopping and some just logging. -camera, transform and mesh components added. -Add and remove mesh functions added to renderer -scene pointer added to every object -3d space and camera is working again

+++ GamEncin/src/Encin/Tools.cpp
@@ -65,7 +65,7 @@ namespace GamEncin
         map<int, KeyButtonStatus> Mouse::buttons;
         Vector2 Mouse::position = Vector2::Zero(),
             Mouse::positionDelta = Vector2::Zero();
-        int Mouse::scrollDelta = 0;
+        float Mouse::scrollDelta = 0;
 
         void Mouse::Initialize(GLFWwindow* window)
         {
